feat: use plugins and framework names
