+++ package.json
@@ -76,8 +76,9 @@
   "dependencies": {
     "@netlify/build": "^9.9.7",
     "@netlify/config": "^4.1.3",
-    "@netlify/framework-info": "^2.3.0",
+    "@netlify/framework-info": "^3.1.0",
     "@netlify/plugin-edge-handlers": "^1.11.5",
+    "@netlify/plugins-list": "^2.2.0",
     "@netlify/traffic-mesh-agent": "^0.27.10",
     "@netlify/zip-it-and-ship-it": "^2.7.1",
     "@oclif/command": "^1.6.1",
